- fixed partitioning bug - renamed Main class of jsettlers.mapcreator - added default close operation to mapcreator - improved response of game by giving LandmarksThreadListener a max speed

+++ src/jsettlers/common/map/shapes/MapCircle.java
@@ -97,7 +97,7 @@ public class MapCircle implements IMapArea {
 		return squaredDistanceToCenter(x, y) < minradius * minradius;
 	}
 
-	public final double getRadius() {
+	public final float getRadius() {
 		return radius;
 	}
